fuse cosmology_clusters -> cosmology
