+++ cloelib/cosmology/cosmology.py
@@ -16,6 +16,7 @@ import jax.numpy as jnp
 
 T = TypeVar("T", bound=Union[jnp.ndarray, np.ndarray])
 
+
 @runtime_checkable
 class Background(Protocol):
 
@@ -25,14 +26,14 @@ class Background(Protocol):
         Hubble parameter at redshift 0 in km s-1 Mpc-1.
         """
         ...
-    
+
     @property
     def h(self) -> float:
         """
         Dimensionless Hubble constant
         """
         ...
-    
+
     @property
     def Omega_b0(self) -> float:
         """
@@ -102,14 +103,14 @@ class Background(Protocol):
         Save internal structure format of possible interface codes
         """
         ...
-    
+
     def Omega_b(self, zs: T) -> T:
         """
         Computes the matter density as a function of redshift.
         """
         ...
 
-    def Omega_m(self, zs: T) -> T:
+    def Omega_m(self, zs: T, nonu: bool) -> T:
         """
         Computes the matter density as a function of redshift.
         """
@@ -138,6 +139,25 @@ class Background(Protocol):
         Calculates the angular diameter distance for given redshifts.
         """
         ...
+
+    def rho_crit(self, zs: T) -> T:
+        """
+        Retrieves the critical density as a function of redshift.
+        """
+        ...
+
+    def dV_dzdO(self, zs: T) -> T:
+        """
+        Volume element per redshit per solid angle.
+        """
+        ...
+
+    def rdrag(self) -> float:
+        """
+        Sound horizon radius at last scattering
+        """
+        ...
+
 
 @runtime_checkable
 class Perturbations(Protocol):
